// apps/desktop/src/bindings/proto/UserRef.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AttachmentId } from "./AttachmentId";
import type { UserId } from "./UserId";

/**
 * Lightweight user reference embedded in messages/member lists.
 */
export type UserRef = { id: UserId, username: string, display_name: string | null, avatar_attachment_id: AttachmentId | null, accent_color: string | null, };